Replace requirements in plug tooltip with stackable label
