+++ src/app/loadout/plug-drawer/PlugStackableIcon.m.scss
@@ -1,0 +1,3 @@
+.stackable {
+  font-size: 11px;
+}
