+++ src/app/loadout/plug-drawer/PlugStackableIcon.m.scss.d.ts
@@ -1,7 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'requirement': string;
+  'stackable': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
