+++ src/app/loadout/plug-drawer/SelectablePlug.m.scss
@@ -5,7 +5,3 @@
   font-size: 11px;
   color: var(--theme-text-secondary);
 }
-
-.stackable {
-  font-size: 11px;
-}
